Module type Hlclock.Timestamp

type t

val walltime : t -> ms

val walltime_unix : t -> float

val to_pair : t -> int * int

val of_pair : (int * int) -> (t, hlc_error) Stdlib.result

val compare : t -> t -> int

compare t t' is a total order on timestamps that is compatible with timeline order.

val equal : t -> t -> bool

equal t t' is true iff t and t' are the same timestamps.

val is_earlier : t -> than:t -> bool

is_earlier t ~than is true iff compare t than = -1.

val is_later : t -> than:t -> bool

is_later t than is true iff compare t than = 1.

val create : ms -> t

val update : ms -> t -> t

val merge : ms:ms -> rt:t -> t -> t

module O : sig ... end

val min : t -> t -> t

val max : t -> t -> t

val to_string : t -> string

val of_string : string -> (t, hlc_error) Stdlib.result

val pp : Stdlib.Format.formatter -> t -> unit